Make Zlib.Compressor/Decompressor classes (#1769)
Motivation:

z_stream stores a pointer to itself in its internal state which it
checks against in inflate/deflate. As we hold these within structs, and
call through to C functions which take a pointer to a z_stream, this
address can change as the struct is copied about. This results in errors
when calling deflate/inflate.

Modifications:

- Hold a pointer to the z_stream

Result:

Harder to misue compressor/decompressor