Make testRetriesCantBeExecutedForTooManyRequestMessages more reliable (#1781)
Motivation:

The 'testRetriesCantBeExecutedForTooManyRequestMessages' test wedges
infrequently. After some diagnosis this appears to be due to the
intersection of two bugs. The first is that when yielding an element to
the sequence, if some consumers exist, of which a subset are waiting for
the next element and the remainder are not 'slow', the continuations of
the fastest consumers would not be resumed when a new element was added.

The other bug is a timing issue: when waiting for the next element a
subscriber may be told to suspend. However, the subscriber must drop and
then reacquire the lock to store the continuation. The state wasn't
re-checked on storing the continuation which opened up a window where
the element may have become present between asking for it and storing
the continuation.

Modifications:

- Don't release the lock between attempting to consume an element and
  suspending to wait for it.
- Resume waiting consumers more frequently.
- Make a few tests less flaky

Result:

Test didn't wedge in 10k iterations